+++ packages/mds-ingest-service/repository/migrations/1660590738992-DropTelemetryAnnotations.ts
@@ -1,6 +1,8 @@
 import type { MigrationInterface, QueryRunner } from 'typeorm'
 
 export class DropTelemetryAnnotations1660590738992 implements MigrationInterface {
+  name = 'DropTelemetryAnnotations1660590738992'
+
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(`DROP TABLE IF EXISTS "telemetry_annotations"`)
   }