+++ packages/mds-policy-service/@types/intents/common.ts
@@ -1,25 +1,27 @@
 import { now } from '@mds-core/mds-utils'
 import type { BaseRule, MicroMobilityStatesToEvents, PolicyDomainModel } from '../models'
 
+export type IntentRuleDefaults = Omit<
+  BaseRule<PartialMicroMobilityStatesMap>,
+  'rule_id' | 'modality' | 'geographies'
+> & {
+  modality: 'micromobility'
+}
 export type PartialMicroMobilityStatesMap = Partial<MicroMobilityStatesToEvents>
 
-export const INTENT_TYPES = <const>['no_parking', 'permitted_vehicle_count', 'parking_time_limit', 'permitted_parking']
+export const INTENT_TYPES = <const>[
+  'no_parking',
+  'permitted_vehicle_count',
+  'parking_time_limit',
+  'permitted_parking',
+  'provider_rebalancing_zone'
+]
 export type INTENT_TYPE = typeof INTENT_TYPES[number]
 
 export const BASE_POLICY_DEFAULTS: Pick<
   PolicyDomainModel,
-  | 'name'
-  | 'description'
-  | 'provider_ids'
-  | 'start_date'
-  | 'end_date'
-  | 'prev_policies'
-  | 'rules'
-  | 'currency'
-  | 'published_date'
+  'provider_ids' | 'start_date' | 'end_date' | 'prev_policies' | 'rules' | 'currency' | 'published_date'
 > = {
-  name: '',
-  description: '',
   provider_ids: [],
   start_date: now(),
   end_date: null,
@@ -36,9 +38,8 @@ export type BaseIntentPolicyUserFields = Pick<
 
 export const COMMON_INTENT_RULE_DEFAULTS: Omit<
   BaseRule<PartialMicroMobilityStatesMap>,
-  'rule_id' | 'maximum' | 'minimum' | 'name' | 'states'
+  'rule_id' | 'maximum' | 'minimum' | 'name' | 'states' | 'geographies'
 > = {
-  geographies: [],
   days: [],
   start_time: null,
   end_time: null,