+++ container-images/mds-activity-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @container-images/mds-activity-service
 
+## 0.1.6
+
+### Patch Changes
+
+- @mds-core/mds-activity-service@0.1.6
+
 ## 0.1.5
 
 ### Patch Changes