+++ container-images/mds-ingest-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @container-images/mds-ingest-service
 
+## 0.2.22
+
+### Patch Changes
+
+- Updated dependencies [9d2b1652]
+  - @mds-core/mds-ingest-service@0.14.1
+
 ## 0.2.21
 
 ### Patch Changes