+++ packages/mds-collector-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mds-core/mds-collector-api
 
+## 0.1.98
+
+### Patch Changes
+
+- @mds-core/mds-collector-service@0.4.44
+- @mds-core/mds-api-helpers@0.3.17
+
 ## 0.1.97
 
 ### Patch Changes